+++ linphone-app/assets/languages/es.ts
@@ -796,6 +796,14 @@ URL del servidor no configurada.</translation>
         <source>byeFunctionDescription</source>
         <translation>Terminar una llamada específica, todas las llamadas o la llamada actual.</translation>
     </message>
+    <message>
+        <source>acceptFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>declineFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>CodecsViewer</name>