+++ linphone-app/assets/languages/ja.ts
@@ -790,6 +790,14 @@
         <source>byeFunctionDescription</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>acceptFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>declineFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>CodecsViewer</name>