+++ linphone-app/assets/languages/lt.ts
@@ -802,6 +802,14 @@ Nesukonfigūruotas serverio url.</translation>
         <source>byeFunctionDescription</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>acceptFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>declineFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>CodecsViewer</name>