+++ linphone-app/assets/languages/pt_BR.ts
@@ -796,6 +796,14 @@ URL do servidor não configurado.</translation>
         <source>byeFunctionDescription</source>
         <translation>Encerre uma chamada específica, todas as chamadas ou a chamada atual.</translation>
     </message>
+    <message>
+        <source>acceptFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>declineFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>CodecsViewer</name>