+++ linphone-app/assets/languages/ru.ts
@@ -802,6 +802,14 @@
         <source>byeFunctionDescription</source>
         <translation>Завершить определённый звонок, все звонки или текущий звонок.</translation>
     </message>
+    <message>
+        <source>acceptFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>declineFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>CodecsViewer</name>