+++ linphone-app/assets/languages/sv.ts
@@ -796,6 +796,14 @@ Serverwebbadressen är inte konfigurerad.</translation>
         <source>byeFunctionDescription</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>acceptFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>declineFunctionDescription</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>CodecsViewer</name>